TST/REF: collect Index formatting tests (#38050)
